fix: move feature operations
